requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
anyio==4.8.0
audioread==3.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
filelock==3.17.0
fsspec==2025.2.0
google-auth==2.38.0
google-genai==1.3.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.65.4
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
psutil==7.0.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.0a2
pydantic_core==2.29.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.2
websockets==14.2
